When doing a Mixmax edit, draw the existing image
Unfortunately, the way this drawing plugin works is by clearing and
redrawing the image based on all the click events it's recorded so far.
This means the existing image gets cleared as soon as you start editing,
so it's not very useful.

This is a limitation of the plugin. We can look for another plugin,
extend this one, or write our own to address the issue